Remove `etc` folder
These files appear to be identical to the ips in the example project. I'll be adding a README with instructions on setting up a project to use this library, which will reference the `ips` in the `examples/blinker` folder. Having one source of truth should be easier to keep updated.